refactor[#190] : 디렉토리 구조 수정. - auth, user 디렉토리로 분리 - user entity 권한 Enumerated 타입 문자열로 변경
