--- com.sparta.logistics.user-service/src/main/java/com/sparta/logistics/user_service/application/dto/request/DeliveryManagerUpdateRequestDto.java
@@ -1,10 +1,0 @@
-package com.sparta.logistics.user_service.application.dto.request;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class DeliveryManagerUpdateRequestDto {
-    private String slackId;
-}
