--- com.sparta.logistics.user-service/src/main/java/com/sparta/logistics/user_service/application/dto/request/UserRoleUpdateRequestDto.java
@@ -1,8 +1,0 @@
-package com.sparta.logistics.user_service.application.dto.request;
-
-import lombok.Getter;
-
-@Getter
-public class UserRoleUpdateRequestDto {
-    private String newRole;
-}
