--- com.sparta.logistics.user-service/src/main/java/com/sparta/logistics/user_service/application/dto/request/UserUpdateRequestDto.java
@@ -1,11 +1,0 @@
-package com.sparta.logistics.user_service.application.dto.request;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class UserUpdateRequestDto {
-    private String newUsername;
-    private String newSlackId;
-}
