--- com.sparta.logistics.user-service/src/main/java/com/sparta/logistics/user_service/application/dto/response/AuthTokenRefreshResponseDto.java
@@ -1,13 +1,0 @@
-package com.sparta.logistics.user_service.application.dto.response;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class AuthTokenRefreshResponseDto {
-
-    private String newAccessToken;
-    private String newRefreshToken;
-
-}
