--- com.sparta.logistics.user-service/src/main/java/com/sparta/logistics/user_service/application/dto/response/UserPasswordResponseDto.java
@@ -1,8 +1,0 @@
-package com.sparta.logistics.user_service.application.dto.response;
-
-import lombok.Data;
-
-@Data
-public class UserPasswordResponseDto {
-
-}
