--- com.sparta.logistics.user-service/src/main/java/com/sparta/logistics/user_service/application/dto/response/UserSearchResponseDto.java
@@ -1,15 +1,0 @@
-package com.sparta.logistics.user_service.application.dto.response;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class UserSearchResponseDto {
-
-    private Long userId;
-    private String username;
-    private String slackId;
-    private String role;
-
-}
