+++ com.sparta.logistics.user-service/src/main/java/com/sparta/logistics/user_service/auth/application/dto/request/AuthLoginRequestDto.java
@@ -1,0 +1,11 @@
+package com.sparta.logistics.user_service.auth.application.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class AuthLoginRequestDto {
+
+    private String username;
+    private String password;
+
+}
