--- com.sparta.logistics.user-service/src/main/java/com/sparta/logistics/user_service/domain/entity/UserRole.java
@@ -1,9 +1,0 @@
-package com.sparta.logistics.user_service.domain.entity;
-
-public enum UserRole {
-    ROLE_MASTER,
-    ROLE_DELIVERY_MANAGER,
-    ROLE_HUB_MANAGER,
-    ROLE_COMPANY_MANAGER,
-    ROLE_USER
-}
