+++ com.sparta.logistics.user-service/src/main/java/com/sparta/logistics/user_service/user/application/dto/request/DeliveryManagerUpdateRequestDto.java
@@ -1,0 +1,10 @@
+package com.sparta.logistics.user_service.user.application.dto.request;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class DeliveryManagerUpdateRequestDto {
+    private String slackId;
+}
