+++ com.sparta.logistics.user-service/src/main/java/com/sparta/logistics/user_service/user/application/dto/request/UserRoleUpdateRequestDto.java
@@ -1,0 +1,8 @@
+package com.sparta.logistics.user_service.user.application.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class UserRoleUpdateRequestDto {
+    private String newRole;
+}
