+++ com.sparta.logistics.user-service/src/main/java/com/sparta/logistics/user_service/user/application/dto/response/UserRoleUpdateResponseDto.java
@@ -1,0 +1,13 @@
+package com.sparta.logistics.user_service.user.application.dto.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class UserRoleUpdateResponseDto {
+
+    private Long userId;
+    private String slackId;
+    private String newRole;
+}
