+++ com.sparta.logistics.user-service/src/main/java/com/sparta/logistics/user_service/user/application/dto/response/UserSearchAllResponseDto.java
@@ -1,0 +1,15 @@
+package com.sparta.logistics.user_service.user.application.dto.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class UserSearchAllResponseDto {
+
+    private Long userId;
+    private String username;
+    private String slackId;
+    private String role;
+
+}
