+++ com.sparta.logistics.user-service/src/main/java/com/sparta/logistics/user_service/user/application/dto/response/UserSearchMeResponseDto.java
@@ -1,0 +1,14 @@
+package com.sparta.logistics.user_service.user.application.dto.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class UserSearchMeResponseDto {
+    private Long id;
+    private String username;
+    private String slackId;
+    private String role;
+
+}
