+++ com.sparta.logistics.user-service/src/main/java/com/sparta/logistics/user_service/user/domain/repository/UserRepository.java
@@ -1,0 +1,20 @@
+package com.sparta.logistics.user_service.user.domain.repository;
+
+import com.sparta.logistics.user_service.user.domain.entity.User;
+import com.sparta.logistics.user_service.user.domain.entity.UserRole;
+import java.util.Optional;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+
+public interface UserRepository {
+
+    Optional<User> findById(Long id);
+
+    void save(User user);
+
+    Optional<User> findByUsername(String username);
+
+    Page<User> findByRole(UserRole roleEnum, Pageable pageable);
+
+    Page<User> findByAll(Pageable pageable);
+}
