+++ com.sparta.logistics.user-service/src/main/java/com/sparta/logistics/user_service/user/infrastructure/jpa/UserJpaRepository.java
@@ -1,0 +1,17 @@
+package com.sparta.logistics.user_service.user.infrastructure.jpa;
+
+import com.sparta.logistics.user_service.user.domain.entity.User;
+import com.sparta.logistics.user_service.user.domain.entity.UserRole;
+import java.util.Optional;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserJpaRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findByUsername(String username);
+
+    Page<User> findByRole(UserRole role, Pageable pageable);
+}
